+++ OniLink/protocol/bedrock-codec/src/main/java/org/cloudburstmc/protocol/bedrock/data/inventory/ContainerType.java
@@ -50,7 +50,9 @@ public enum ContainerType {
     /**
      * @since v630
      */
-    CRAFTER(36);
+    CRAFTER(36),
+    /** @since v2192 */
+    DATA_DRIVEN_CONTAINER(37);
 
     public static final ContainerType[] VALUES;